# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: MCM AI Lookbook Kiosk API
  version: 0.2.0
  summary: Kiosk, recommendation, product, conversion and manager API contract
  description: >-
    Contract-first scaffold for parallel FE, BE, Eye and Face development.
    Raw webcam frames are not accepted by any operation.
servers:
  - url: /
    description: Same-origin development server
...
paths:
  /api/v1/sessions:
    post:
      tags: [Sessions]
      operationId: createSession
      summary: Create an anonymous kiosk session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/sessions/{session_id}/vision-stream-token:
    post:
      tags: [Sessions]
      operationId: issueVisionStreamToken
      summary: Issue a short-lived one-time Vision Stream credential
      description: >-
        The token is bound to the active, consented session and its video. It
        authorizes only the ephemeral binary WSS Vision Stream and must not be
        logged, persisted or placed in the WebSocket URL.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: One-time Vision Stream credential
          content:
            application/json:
              schema:
                $ref: "./vision-stream-v1/vision-stream-token.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/lookbooks/{lookbook_id}/manifest:
    get:
      tags: [Lookbooks]
      operationId: getLookbookManifest
      summary: Get the versioned video and AOI manifest
      parameters:
        - $ref: "#/components/parameters/LookbookId"
      responses:
        "200":
          description: Lookbook manifest
          content:
            application/json:
              schema:
                $ref: "./lookbook-manifest.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/sessions/{session_id}/reaction-batches:
    post:
      tags: [Reactions]
      operationId: appendReactionBatch
      summary: Accept and immediately aggregate a batch of derived reaction events
      description: >-
        The backend validates and de-duplicates an active session's batch, then
        folds valid signals into privacy-minimized product features. Individual
        event payloads are not persisted in the MVP C profile.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./events/reaction-batch.schema.json"
      responses:
        "202":
          description: Batch accepted or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionBatchAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/sessions/{session_id}/complete:
    post:
      tags: [Recommendations]
      operationId: completeSessionAnalysis
      summary: Finish signal ingestion and request a recommendation
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "202":
          description: Recommendation processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/sessions/{session_id}/recommendations:
    get:
      tags: [Recommendations]
      operationId: getSessionRecommendation
      summary: Get the current Top 2 recommendation state
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Recommendation state
          content:
            application/json:
              schema:
                $ref: "./events/recommendation-result.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/sessions/{session_id}/manager-product-requests:
    post:
      tags: [Managers]
      operationId: createManagerProductRequest
      summary: Request manager assistance for a completed Top 2 recommendation
      description: >-
        The request body accepts only request_id and recommendation_id. Backend
        verifies that the recommendation belongs to the session and is completed,
        then derives the Top 2 items for the manager event from the stored result.
        Repeating the same request_id is handled idempotently.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./requests/manager-product-request.schema.json"
      responses:
        "202":
          description: Customer product request accepted or recognized as a duplicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerProductRequestAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/products/{product_id}:
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get product image, official URL and pre-generated QR metadata
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/conversions:
    post:
      tags: [Conversions]
      operationId: recordConversionOutcome
      summary: Record a manager-entered fitting or purchase outcome
      description: A retry with the same outcome_id must be an identical document; reuse with different values returns a conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./events/conversion-outcome.schema.json"
      responses:
        "201":
          description: Conversion outcome recorded
          content:
            application/json:
              schema:
                $ref: "./events/conversion-outcome.schema.json"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/manager/events:
    get:
      tags: [Managers]
      operationId: listManagerEvents
      summary: Poll customer product request events after a sequence cursor
      description: Manager clients poll this endpoint every 1-2 seconds and advance the cursor with the largest returned sequence.
      parameters:
        - name: after_sequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Ordered manager events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./events/manager-event.schema.json"
  /api/v2/sessions/{session_id}/observations:
    post:
      tags: [Reactions]
      operationId: appendObservationBatchV2
      summary: Accept a batch of derived frame observations for the central recommender
      description: >-
        Accepts only privacy-minimized derived signals whose video_time_ms was
        snapshotted when the camera frame was captured. Raw frames, image bytes,
        base64/data URIs, embeddings, landmarks and original/source paths are forbidden.
        Kiosk producers must send an empty attention.candidates array; Backend rejects
        client product attribution and maps only approved server-side AOI metadata.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./v2/observation-batch-v2.schema.json"
      responses:
        "202":
          description: Observation batch accepted or already processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservationBatchAcceptedV2"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          description: Observation batch exceeds the configured payload limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v2/sessions/{session_id}/complete:
    post:
      tags: [Recommendations]
      operationId: completeSessionAnalysisV2
      summary: Freeze derived observations and enqueue a central Top 1 decision
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "202":
          description: Central recommendation processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationAcceptedV2"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v2/sessions/{session_id}/recommendation:
    get:
      tags: [Recommendations]
      operationId: getSessionRecommendationV2
      summary: Poll the pending receipt or terminal central Top 1 decision
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Terminal central recommendation decision
          content:
            application/json:
              schema:
                $ref: "./v2/recommendation-decision-v2.schema.json"
        "202":
          description: Central recommendation is still pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationAcceptedV2"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v2/sessions/{session_id}:
    delete:
      tags: [Sessions]
      operationId: deleteSessionV2
      summary: Cancel the session and discard all session-local derived evidence
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          description: Session-local observations, evidence and de-duplication keys discarded
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v2/products/{product_id}:
    get:
      tags: [Products]
      operationId: getProductV2
      summary: Get one reviewed central recommendation product profile
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product profile with explicit nullable URL, image and QR provenance
          content:
            application/json:
              schema:
                $ref: "./v2/product-detail-v2.schema.json"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v2/sessions/{session_id}/manager-product-requests:
    post:
      tags: [Managers]
      operationId: createManagerProductRequestV2
      summary: Send an explicit customer request for the completed Top 1 product
      description: >-
        Backend verifies that recommendation_id belongs to the session and that
        selected_product_id exactly matches its completed decision. No event is
        emitted automatically when a session starts or a recommendation completes.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./v2/manager-product-request-v2.schema.json"
      responses:
        "202":
          description: Customer product request accepted or recognized as a duplicate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerProductRequestAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v2/manager/events:
    get:
      tags: [Managers]
      operationId: listManagerEventsV2
      summary: Poll explicit customer Top 1 product request events
      parameters:
        - name: after_sequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        "200":
          description: Ordered customer product request events only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./v2/manager-event-v2.schema.json"
...
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    LookbookId:
      name: lookbook_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    ProductId:
      name: product_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
  schemas:
    Identifier:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]*$"
    SessionCreate:
      type: object
      additionalProperties: false
      required: [kiosk_id, lookbook_id, consent_version]
      properties:
        kiosk_id:
          $ref: "#/components/schemas/Identifier"
        lookbook_id:
          $ref: "#/components/schemas/Identifier"
        consent_version:
          $ref: "#/components/schemas/Identifier"
    SessionCreated:
      type: object
      additionalProperties: false
      required: [session_id, display_code, status, created_at]
      properties:
        session_id:
          $ref: "#/components/schemas/Identifier"
        display_code:
          $ref: "#/components/schemas/Identifier"
        status:
          enum: [created, collecting]
        created_at:
          type: string
          format: date-time
    ReactionBatchAccepted:
      type: object
      additionalProperties: false
      required: [batch_id, status]
      properties:
        batch_id:
          $ref: "#/components/schemas/Identifier"
        status:
          enum: [accepted, duplicate]
    RecommendationAccepted:
      type: object
      additionalProperties: false
      required: [session_id, status]
      properties:
        session_id:
          $ref: "#/components/schemas/Identifier"
        status:
          const: pending
    ObservationBatchAcceptedV2:
      type: object
      additionalProperties: false
      required: [batch_id, status]
      properties:
        batch_id:
          $ref: "#/components/schemas/Identifier"
        status:
          enum: [accepted, duplicate]
    RecommendationAcceptedV2:
      type: object
      additionalProperties: false
      required: [recommendation_id, decision_request_id, status]
      properties:
        recommendation_id:
          $ref: "#/components/schemas/Identifier"
        decision_request_id:
          $ref: "#/components/schemas/Identifier"
        status:
          const: pending
    ManagerProductRequestAccepted:
      type: object
      additionalProperties: false
      required: [request_id, status]
      properties:
        request_id:
          $ref: "#/components/schemas/Identifier"
        status:
          enum: [accepted, duplicate]
    Product:
      type: object
      additionalProperties: false
      required: [product_id, display_name, category, image_url, product_url, qr_asset_path]
      properties:
        product_id:
          $ref: "#/components/schemas/Identifier"
        display_name:
          type: string
          minLength: 1
          maxLength: 200
        category:
          type: string
          minLength: 1
          maxLength: 100
        image_url:
          type: string
          format: uri-reference
        product_url:
          type: string
          format: uri
        qr_asset_path:
          type: string
          minLength: 1
          maxLength: 500
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 500
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Session state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"